complete parser day 2

+++ Lesson3/Day2/Solution/Parser/Parser/Source code/scanner.c
@@ -88,7 +88,7 @@ Token* readNumber(void) {
         readChar();
         i++;
     }
-    if (charCodes[token->string[i-1]] != CHAR_DIGIT) {
+    if (charCodes[(int)token->string[i-1]] != CHAR_DIGIT) {
         error(ERR_INVALIDFLOATNUMBER, lineNo, colNo);
     }
     
@@ -98,7 +98,8 @@ Token* readNumber(void) {
     if (numDot == 1) {
         token->tokenType = TK_FLOAT;
     }
-    return token;}
+    return token;
+}
 
 Token* readConstChar(void) {
     Token *token = makeToken(TK_CHAR, lineNo, colNo);
